Add zfbfinxmin extension (#31)
With Zfinx, Zdinx, Zhinx, and Zhinxmin there's a clear precedent that
floating point extensions should have an analogous "inx" version. As
defining such a variant is straight-forward, I feel it would be most
sensible (and efficient) to define it as part of the bfloat16 definition
process rather than leaving a gap within the ISA to be filled at a later
date.

This drafts such a definition. As the bfloat16 extensions are defined in
new documentation rather than patches to the riscv-isa-manual, I've
duplicated some explanatory text from the Zfinx chapater of the manual
so that the zfbfinxmin definition makes sense standalone. If merging
into the ISA manual, it would of course take a different form.